test(saves): keep the properties the vectors cannot express
The save-sync decision runs in the compiled gavel core, so the hand-enumerated
decision cases and the property tier now both decide through
GavelNativeAdapter and live under tests/adapters/, next to the code they cover.

Seven properties restated a single row of the decision table — a point
statement binding one input to one output, which is exactly what the vendored
gavel vectors already carry. That is two places to maintain for one rule, so
they are gone. The five that stay either quantify over a space the vectors do
not exhaust, or relate several runs to each other (two ISO renderings of the
same instant, a replay, a state sequence) — something a vector cannot express
at all.

The byte-identity property no longer samples shrunken locals. The branch-4
corrupt-local guard is thereby out of its search space, so every Conflict is a
failure with no case distinction, and the shrink threshold stays in the single
place that owns it.
